moved base from algorithm store to common
