+++ vantage6-algorithm-store/vantage6/algorithm/store/model/algorithm.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, String
 from sqlalchemy.orm import relationship
 
-from vantage6.algorithm.store.model.base import Base
+from vantage6.common.base import Base
 
 
 class Algorithm(Base):
