+++ vantage6-algorithm-store/vantage6/algorithm/store/model/permission.py
@@ -11,7 +11,7 @@ may contain multiple entries for the same user.
 
 from sqlalchemy import Column, Integer, ForeignKey, Table
 
-from .base import Base
+from vantage6.common.base import Base
 
 
 Permission = Table(
