+++ vantage6-algorithm-store/vantage6/algorithm/store/model/role_rule_association.py
@@ -6,7 +6,7 @@ role will usually have multiple rules assigned to it.
 
 from sqlalchemy import Column, Integer, ForeignKey, Table
 
-from .base import Base
+from vantage6.common.base import Base
 
 # TODO: refactor to use the server model
 role_rule_association = Table(
