+++ vantage6-algorithm-store/vantage6/algorithm/store/permission.py
@@ -6,10 +6,9 @@ from flask_principal import Permission
 
 from vantage6.algorithm.store.globals import RESOURCES
 from vantage6.algorithm.store.default_roles import DefaultRole
-from vantage6.algorithm.store.model.base import Base
 from vantage6.algorithm.store.model.role import Role
 from vantage6.algorithm.store.model.rule import Rule, Operation
-from vantage6.algorithm.store.model.base import DatabaseSessionManager
+from vantage6.common.base import DatabaseSessionManager
 from vantage6.common import logger_name
 
 module_name = logger_name(__name__)
